Deduplicated integer conversion

+++ src/gameplay/cdda/map_memory_manager.rs
@@ -31,8 +31,8 @@ impl<'w> MapMemoryManager<'w> {
                 .handle(&self.asset_server, &self.assets, base_zone_level, path);
         match map_memory {
             AssetState::Available { asset: map_memory } => {
-                let index = subzone_level.z.rem_euclid(8) as usize * 8
-                    + subzone_level.x.rem_euclid(8) as usize;
+                let index =
+                    (subzone_level.z.rem_euclid(8) * 8 + subzone_level.x.rem_euclid(8)) as usize;
                 AssetState::Available {
                     asset: &map_memory.0[index],
                 }
